+++ src/components/LayerCollection/LayerPOLine.js
@@ -9,6 +9,7 @@ import {
 import ReactRouterPropTypes from 'react-router-prop-types';
 import queryString from 'query-string';
 
+import { stripesShape } from '@folio/stripes/core';
 import {
   Layer,
 } from '@folio/stripes/components';
@@ -39,7 +40,6 @@ import {
 } from '../Utils/resources';
 import { POLineForm } from '../POLine';
 import LinesLimit from '../PurchaseOrder/LinesLimit';
-import { DEFAULT_CURRENCY } from '../POLine/Cost/FieldCurrency';
 import getOrderTemplateValue from '../Utils/getOrderTemplateValue';
 
 const ERROR_CODES = {
@@ -216,7 +216,7 @@ class LayerPOLine extends Component {
       template: get(order, 'template', ''),
       source: sourceValues.user,
       cost: {
-        currency: get(vendor, 'vendorCurrencies[0]', DEFAULT_CURRENCY),
+        currency: stripes.currency,
       },
       vendorDetail: {
         instructions: '',
@@ -356,10 +356,7 @@ LayerPOLine.propTypes = {
   parentResources: PropTypes.object.isRequired,
   resources: PropTypes.object.isRequired,
   showToast: PropTypes.func.isRequired,
-  stripes: PropTypes.shape({
-    store: PropTypes.object.isRequired,
-    connect: PropTypes.func.isRequired,
-  }).isRequired,
+  stripes: stripesShape.isRequired,
   onCancel: PropTypes.func.isRequired,
   mutator: PropTypes.object.isRequired,
 };